refactor: resolver "constraints"
