+++ src/lib/prisma-utils/index.ts
@@ -1,0 +1,1 @@
+export * as PrismaUtils from './index_';
